rename: LocalStorage.js를 util/위치로 이동
